+++ internal/provider/provider.go
@@ -115,7 +115,6 @@ func Provider() *schema.Provider {
 			"tfe_organization_tags":       dataSourceTFEOrganizationTags(),
 			"tfe_slug":                    dataSourceTFESlug(),
 			"tfe_ssh_key":                 dataSourceTFESSHKey(),
-			"tfe_team":                    dataSourceTFETeam(),
 			"tfe_teams":                   dataSourceTFETeams(),
 			"tfe_team_access":             dataSourceTFETeamAccess(),
 			"tfe_team_project_access":     dataSourceTFETeamProjectAccess(),